Fix row swap error on matrices with duplicate rows
If we moved the leading value inward by two and had no rows with it in
the expected position, the program would break. To fix this, I removed
hardcoded references to the leading position as matrix[i][i] and also
kept track of the leading position of the last row to compare to.